Add files via upload
This commit adds in the recipe menu and pantry menu to the main menu.  The buttons that send you to these menus are now active.  The only button left inactive is the login.  More info available in README.txt that I will be uploading to the branch by the end of Monday.